requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
python-nmap==0.7.1
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
